renamed nonce_generator.rb --> NonceGenerator.rb
