+++ lib/bitstampede/NonceGenerator.rb
